refactor: 뼈대 코드 수정

+++ src/main/java/cart/dto/ProductRequest.java
@@ -1,14 +1,16 @@
 package cart.dto;
 
+import java.math.BigDecimal;
+
 public class ProductRequest {
     private String name;
-    private int price;
+    private BigDecimal price;
     private String imageUrl;
 
     public ProductRequest() {
     }
 
-    public ProductRequest(String name, int price, String imageUrl) {
+    public ProductRequest(String name, BigDecimal price, String imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
@@ -18,7 +20,7 @@ public class ProductRequest {
         return name;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
